Split out NumberParts
